Изменение контроллера согласно: https://docs.microsoft.com/ru-ru/azure/architecture/best-practices/api-design

+++ backend/src/main/java/application/services/ClientsService.java
@@ -1,5 +1,6 @@
 package application.services;
 
+import application.exception.ResourceNotFoundException;
 import application.models.Client;
 import application.repositories.ClientRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,8 +25,15 @@ public class ClientsService {
         return clientRepository.findAll();
     }
 
-    public void delete(Client client) {
+    public ResponseEntity<?> delete(Long id) {
+        Client client = clientRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Client with id:" + id + " not found.")
+        );
+        if (client == null) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         clientRepository.delete(client);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     public Optional<Client> getById(Long id) {
